+++ src/RubySettings.cpp
@@ -3,6 +3,7 @@
 void RubySettings::toJson(JsonDocument& doc) const {
   doc["wifiSniffEnabled"] = wifiSniffEnabled;
   doc["bleSniffEnabled"] = bleSniffEnabled;
+  doc["rawHandshakeCaptureEnabled"] = rawHandshakeCaptureEnabled;
   doc["wifiChannelDwellMs"] = wifiChannelDwellMs;
   doc["clockUtcOffsetQ"] = clockUtcOffsetQ;
   doc["fullRefreshIntervalMin"] = fullRefreshIntervalMin;
@@ -11,6 +12,7 @@ void RubySettings::toJson(JsonDocument& doc) const {
 bool RubySettings::fromJson(JsonVariantConst doc) {
   wifiSniffEnabled = doc["wifiSniffEnabled"] | true;
   bleSniffEnabled = doc["bleSniffEnabled"] | true;
+  rawHandshakeCaptureEnabled = doc["rawHandshakeCaptureEnabled"] | false;
   wifiChannelDwellMs = doc["wifiChannelDwellMs"] | 300;
   clockUtcOffsetQ = doc["clockUtcOffsetQ"] | 48;
   fullRefreshIntervalMin = doc["fullRefreshIntervalMin"] | 20;